changing detail poMatch to Float type

+++ app.py
@@ -28,8 +28,9 @@ detail = pd.read_csv(detail_file)
 
 #### Merge()  VLOOKUP Equivalent
 
-remit['poMatch']  = remit['Invoice Number'].astype(float)
-detail['poMatch']  = detail['Purchase Order Number'].astype(float)
+remit['poMatch']  = remit['Invoice Number']
+detail['poMatch']  = detail['Purchase Order Number']
+detail['poMatch'] = detail['poMatch'].astype(float)
 remit_review = remit[['poMatch','Amount Paid($)','Store Number','Invoice Date']]
 detail_review = detail[['poMatch','Ship-To Customer Name','Invoice Number','Requested Delivery Date']]
